New autolocate / WKD Lookup for key discovery
This renames a the tofu / keyserver lookup to a generic
autolocate option, which is enabled if one additional source
is used.

The current additional sources are the Mailvelope Keyserver
and Web Key Directory (WKD). The AutoLocate class is a small
wrapper to provide an API to query both.

Details:
* src/modules/AutoLocate.js: New module to provide an API to
query additional key discovery sources. The name is based on
GnuPG's auto-key-locate options.

* src/modules/WKDLocate.js: Implementation of WKD Lookup.

* src/components/editor/components/RecipientInput.js,
src/components/editor/editor.js: Wording / names
changed to autolocate.

* src/controller/editor.controller.js: Use autolocate instead
of using the keyserver directly.